deleted formatted. now in robocup-actions
